Clean up the GitCloneView
